% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composeBox.R
\name{composeBoxUI}
\alias{composeBoxUI}
\alias{composeBox}
\title{A Shiny module for composing messages}
\usage{
composeBoxUI(
  id,
  labels = list(bericht = "Bericht", tag_gebruiker = "Notificatie gebruiker", verstuur
    = "Verstuur"),
  users = NULL
)

composeBox(input, output, session, msg, attachment = "", current_user = NULL)
}
\arguments{
\item{id}{A unique ID for the module instance.}

\item{labels}{A list with labels.}

\item{users}{Optionally, a vector with user names. If provided, the compose box will include an option 
to 'tag' users for the current message.}
}
\description{
A module (consisting of a UI and a server function) for composing messages.
}
