% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyinbox.R
\name{shinyinboxUI}
\alias{shinyinboxUI}
\alias{shinyinbox}
\title{A Shiny module for an inbox with messages}
\usage{
shinyinboxUI(
  id,
  label_inbox = getOption("sib_inbox", "Inbox"),
  label_delete = getOption("sib_delete", "Verwijder selectie")
)

shinyinbox(
  input,
  output,
  session,
  msg,
  filter_attachment = NULL,
  filter_user = NULL,
  attachment_function = NULL,
  label_message = getOption("sib_message", "Bericht"),
  label_undo = getOption("sib_undo", "Undo"),
  label_tagged = getOption("sib_tagged", "Tagged"),
  label_close = getOption("sib_close", "Sluiten"),
  label_edit = getOption("sib_edit", "Edit"),
  label_save = getOption("sib_save", "Opslaan")
)
}
\arguments{
\item{id}{A unique ID for the inbox UI.}

\item{input}{For the server function, Shiny usage.}

\item{output}{For the server function, Shiny usage.}

\item{session}{For the server function, Shiny usage.}

\item{msg}{A configuration object.}

\item{filter_attachment}{If set, only shows messages that include this value in the attachment.}

\item{language}{A list of labels to be used in the UI.}
}
\description{
Inbox for messages stored in a local database with edit and delete functionality. 
For use in Shiny apps.
}
\examples{
# The configuration object may look like:
msg <- list(
  connection = dbConnect(RSQLite::SQLite(), "data/messages.sqlite"),
  table = "messages",
  edit_rights = TRUE,
  delete_rights = TRUE,
  poll_delay = 500)
}
