% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyinbox.R
\name{shinyinboxUI}
\alias{shinyinboxUI}
\alias{shinyinbox}
\title{A Shiny module for an inbox with messages}
\usage{
shinyinboxUI(
  id,
  language = list(tab_inbox = "Inbox", tab_message = "Bericht", tab_edit = "Edit",
    btn_delete = "Verwijder selectie", btn_close = "Sluiten", btn_save = "Opslaan",
    btn_undo = "Undo", txt_message = "Bericht")
)

shinyinbox(
  input,
  output,
  session,
  msg,
  filter_attachment = NULL,
  filter_user = NULL,
  attachment_function = NULL
)
}
\arguments{
\item{id}{A unique ID for the inbox UI.}

\item{language}{A list of labels to be used in the UI.}

\item{input}{For the server function, Shiny usage.}

\item{output}{For the server function, Shiny usage.}

\item{session}{For the server function, Shiny usage.}

\item{msg}{A configuration object.}

\item{filter_attachment}{If set, only shows messages that include this value in the attachment.}
}
\description{
Inbox for messages stored in a local database with edit and delete functionality. 
For use in Shiny apps.
}
\examples{
# The configuration object may look like:
msg <- list(
  connection = dbConnect(RSQLite::SQLite(), "data/messages.sqlite"),
  table = "messages",
  edit_rights = TRUE,
  delete_rights = TRUE,
  poll_delay = 500)
}
